+++ src/ditool/part.h
@@ -21,14 +21,14 @@ enum Error {
 
 struct part_t {
     int                      partNo;
-    size_t                   partIdx;
+    int                      partIdx;
     struct im_t*             im;
     const struct disk_label* dl;
     struct disk_partition*   part;
     struct part_t*           next;
 };
 
-void partition_init(int partNo, size_t partIdx, struct im_t* im, const struct disk_label* dl, struct disk_partition* partition);
+void partition_init(int partNo, int partIdx, struct im_t* im, const struct disk_label* dl, struct disk_partition* partition);
 void partition_uninit(struct part_t* part);
 int  partition_readSectors(struct part_t* part, uint32_t sector, uint32_t count, uint8_t* dst);
 void partition_print(struct part_t* part);
